Corrections for width of loop count variable.

+++ Common/DtaHashPwd.cpp
@@ -115,7 +115,7 @@ int Testsedutil(const PBKDF_TestTuple *testSet, unsigned int testSetSize)
         const PBKDF_TestTuple &tuple = testSet[i];
         hash.clear();
         seaSalt.clear();
-        for (uint16_t j = 0; j < strnlen(tuple.Salt, 255); j++) {
+        for (size_t j = 0; j < strnlen(tuple.Salt, 255); j++) {
             seaSalt.push_back(tuple.Salt[j]);
         }
 		printf("Password %s Salt %s Iterations %i Length %i\n", (char *)tuple.Password,
